+++ fe/fe-core/src/main/java/com/starrocks/catalog/TabletStatMgr.java
@@ -47,6 +47,7 @@ import com.starrocks.thrift.TTabletStatResult;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ExecutionException;
@@ -60,11 +61,17 @@ public class TabletStatMgr extends LeaderDaemon {
     private static final Logger LOG = LogManager.getLogger(TabletStatMgr.class);
 
     // for lake table
-    private Map<Long, Long> partitionToUpdatedVersion;
+    private final Map<Long, Long> partitionToUpdatedVersion;
+
+    private LocalDateTime lastWorkTimestamp;
 
     public TabletStatMgr() {
         super("tablet stat mgr", Config.tablet_stat_update_interval_second * 1000L);
         partitionToUpdatedVersion = Maps.newHashMap();
+    }
+
+    public LocalDateTime getLastWorkTimestamp() {
+        return lastWorkTimestamp;
     }
 
     @Override
@@ -107,6 +114,7 @@ public class TabletStatMgr extends LeaderDaemon {
         }
         LOG.info("finished to update index row num of all databases. cost: {} ms",
                 (System.currentTimeMillis() - start));
+        lastWorkTimestamp = LocalDateTime.now();
     }
 
     private void updateLocalTabletStat() {
